+++ src/main/java/coffeeshop/entity/Orders.java
@@ -1,6 +1,7 @@
 package coffeeshop.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import javax.persistence.Basic;
@@ -42,12 +43,10 @@ public class Orders implements Serializable {
     @Column(name = "odate")
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime dateCreated;
-//    @Max(value=999.99)  
-//    @Min(value=000.00)//if you know range of your decimal fields consider using these annotations to enforce field validation
     @DecimalMin("000.00")
     @DecimalMax("999.00") 
     @Column(name = "orderprice", precision=5, scale=2)
-    private double price;
+    private BigDecimal price;
     @Size(max = 150)
     @Column(name = "comments")
     private String comments;
@@ -90,11 +89,11 @@ public class Orders implements Serializable {
         this.dateCreated = dateCreated;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
